Move to same logging logic as carbonzipper
As the logging levels were slightly different, I translated them as:
    Warn/Error -> Log
    Info/Debug -> Debug

The two instance of calls to log.Fatal() are left as-is, which means
their errors will go to the stderr log instead of where the rest of the
logging output goes.  However, since these calls are rare and will
execute immediately upon startup, I consider this "ok" that they're not
in the same place as everything else.